# code/iginx-assoc-backend/src/main/java/com/xmu/iginx/assoc/modules/external/vo/ExternalErrorResponse.java
package com.xmu.iginx.assoc.modules.external.vo;

import lombok.Data;

@Data
public class ExternalErrorResponse {

    private String code;

    private String message;

    private String traceId;
}