# code/iginx-assoc-backend/src/main/java/com/xmu/iginx/assoc/modules/external/vo/ExternalJobCreateResponse.java
package com.xmu.iginx.assoc.modules.external.vo;

import lombok.Data;

import java.time.LocalDateTime;

@Data
public class ExternalJobCreateResponse {

    private String jobId;

    private String status;

    private LocalDateTime submitTime;
}